fix: harden display recovery and forced redraw handling
Upstream commits applied:

- f5945344cc21ebff07eb7ea04dab0555fbe49007 gpui(windows): Fix unwrap panic when monitor goes missing (#55630)

- 7d19e899889fc2e4a7999e9ce77a4bd438e3dc3b Fix DirectX atlas panic after GPU device recovery (#55878)

- 38270bc027a46671f680b09d06cd7a13a62bbc99 Fix Alt modifier stuck after Alt-Tab on Windows (#52220)

Review: widen DisplayId to carry raw platform IDs, fall back safely when a monitor handle cannot be resolved, force refresh on forced draws, and resync Windows modifiers on activation.

Verification: cargo test -p adabraka-gpui --lib --features test-support

+++ crates/gpui/src/platform/linux/wayland/window.rs
@@ -471,7 +471,7 @@ fn output_for_layer_shell(
     params.display_id.and_then(|display_id| {
         outputs
             .iter()
-            .find(|output| output.id.protocol_id() == display_id.0)
+            .find(|output| output.id.protocol_id() as u64 == display_id.0)
             .map(|output| output.output.clone())
     })
 }
@@ -569,7 +569,7 @@ fn find_output_for_params<'a>(params: &WindowParams, outputs: &'a [Output]) -> O
         .and_then(|display_id| {
             outputs
                 .iter()
-                .find(|output| output.id.protocol_id() == display_id.0)
+                .find(|output| output.id.protocol_id() as u64 == display_id.0)
         })
         .or_else(|| {
             outputs.iter().find(|output| {
